make test pretty
